requirements.txt
torch==2.4.1
torchaudio==2.4.1
coqui-tts
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydub==0.25.1